+++ tests/handlers/workflow_factory_test.py
@@ -6,7 +6,7 @@ from ess.reduce.streaming import StreamProcessor
 from pydantic import BaseModel, ValidationError
 
 from beamlime.config.workflow_spec import WorkflowConfig, WorkflowSpec
-from beamlime.handlers.stream_processor_factory import WorkflowFactory
+from beamlime.handlers.workflow_factory import WorkflowFactory
 
 
 class MyParams(BaseModel):
@@ -14,7 +14,7 @@ class MyParams(BaseModel):
     name: str = "test"
 
 
-def make_dummy_stream_processor() -> StreamProcessor:
+def make_dummy_workflow() -> StreamProcessor:
     """Fixture to create a mock StreamProcessor."""
     workflow = sciline.Pipeline()
     return StreamProcessor(
@@ -22,7 +22,7 @@ def make_dummy_stream_processor() -> StreamProcessor:
     )
 
 
-def make_dummy_stream_processor_with_source(*, source_name: str) -> StreamProcessor:
+def make_dummy_workflow_with_source(*, source_name: str) -> StreamProcessor:
     """Fixture to create a mock StreamProcessor that uses source_name."""
     workflow = sciline.Pipeline()
     # In a real implementation, the source_name would be used to customize the processor
@@ -31,7 +31,7 @@ def make_dummy_stream_processor_with_source(*, source_name: str) -> StreamProces
     )
 
 
-def make_dummy_stream_processor_with_params(*, params: MyParams) -> StreamProcessor:
+def make_dummy_workflow_with_params(*, params: MyParams) -> StreamProcessor:
     """Fixture to create a mock StreamProcessor that uses params."""
     workflow = sciline.Pipeline()
     return StreamProcessor(
@@ -39,7 +39,7 @@ def make_dummy_stream_processor_with_params(*, params: MyParams) -> StreamProces
     )
 
 
-class TestStreamProcessorFactory:
+class TestWorkflowFactory:
     def test_init_factory_is_empty(self):
         factory = WorkflowFactory()
         assert len(factory) == 0
@@ -59,7 +59,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         assert len(factory) == 1
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
@@ -85,7 +85,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         stored_spec = factory[workflow_id]
@@ -105,14 +105,14 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         with pytest.raises(
             ValueError,
             match="Workflow ID 'test-instrument/test-namespace/test-wf/1' is already "
             "registered.",
         ):
-            factory.register(spec)(lambda: make_dummy_stream_processor())
+            factory.register(spec)(lambda: make_dummy_workflow())
 
     def test_create_returns_stream_processor(self):
         factory = WorkflowFactory()
@@ -128,7 +128,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(identifier=workflow_id)
@@ -150,7 +150,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func(*, source_name):
-            return make_dummy_stream_processor_with_source(source_name=source_name)
+            return make_dummy_workflow_with_source(source_name=source_name)
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(identifier=workflow_id)
@@ -171,7 +171,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func(*, params: MyParams):
-            return make_dummy_stream_processor_with_params(params=params)
+            return make_dummy_workflow_with_params(params=params)
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(
@@ -194,7 +194,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func(*, params: MyParams):
-            return make_dummy_stream_processor_with_params(params=params)
+            return make_dummy_workflow_with_params(params=params)
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(
@@ -228,7 +228,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(identifier=workflow_id)
@@ -259,11 +259,11 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec1)
         def factory_func1():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         @factory.register(spec2)
         def factory_func2():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         assert len(factory) == 2
         specs = list(factory.values())
@@ -284,7 +284,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace/workflow1/1"
         assert factory[workflow_id].name == "workflow1"
@@ -317,11 +317,11 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec1)
         def factory_func1():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         @factory.register(spec2)
         def factory_func2():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         # Both functions should be registered with different IDs but same name
         assert len(factory) == 2
@@ -358,7 +358,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace//1"
         stored_spec = factory[workflow_id]
@@ -385,7 +385,7 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec)
         def factory_func():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         workflow_id = "test-instrument/test-namespace/test-workflow/1"
         config = WorkflowConfig(identifier=workflow_id)
@@ -428,11 +428,11 @@ class TestStreamProcessorFactory:
 
         @factory.register(spec1)
         def factory_func1():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         @factory.register(spec2)
         def factory_func2():
-            return make_dummy_stream_processor()
+            return make_dummy_workflow()
 
         expected_sources = {"source1", "source2", "source3"}
         assert factory.source_names == expected_sources
